third_party/blink/web_tests/platform/generic/fast/css/getComputedStyle/getComputedStyle-background-shorthand-expected.txt

Tests that the background shorthand is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('background') is "rgb(255, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box"
PASS checkComputedStyleValue() is true
PASS computedStyle.getPropertyValue('background') is "rgb(255, 0, 0) url(\"dummy://test.png\") repeat scroll 0% 0% / auto padding-box border-box"
PASS checkComputedStyleValue() is true
PASS computedStyle.getPropertyValue('background') is "rgb(255, 0, 0) url(\"dummy://test.png\") no-repeat scroll 0% 0% / auto padding-box border-box"
PASS checkComputedStyleValue() is true
PASS computedStyle.getPropertyValue('background') is "rgb(255, 0, 0) url(\"dummy://test.png\") no-repeat fixed 0% 0% / auto padding-box border-box"
PASS checkComputedStyleValue() is true
PASS computedStyle.getPropertyValue('background') is "rgb(255, 0, 0) url(\"dummy://test.png\") no-repeat fixed 100% 100% / auto padding-box border-box"
PASS checkComputedStyleValue() is true
PASS computedStyle.getPropertyValue('background') is "rgb(255, 0, 0) url(\"dummy://test.png\") no-repeat fixed 100% 100% / cover padding-box border-box"
PASS checkComputedStyleValue() is true
PASS computedStyle.getPropertyValue('background') is "rgb(255, 0, 0) url(\"dummy://test.png\") no-repeat fixed 100% 100% / cover content-box border-box"
PASS checkComputedStyleValue() is true
PASS computedStyle.getPropertyValue('background') is "rgb(255, 0, 0) url(\"dummy://test.png\") no-repeat fixed 100% 100% / cover content-box padding-box"
PASS checkComputedStyleValue() is true
PASS computedStyle.getPropertyValue('background') is "rgb(0, 128, 0) url(\"dummy://test.png\") repeat fixed 45% 50% / contain border-box padding-box"
PASS checkComputedStyleValue() is true
PASS successfullyParsed is true

TEST COMPLETE
